requirements.txt
-f https://download.pytorch.org/whl/cu126
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
transformers
datasets
pandas
tqdm
scikit-learn
opencv-python
matplotlib
Pillow
albumentations
yacs
pyyaml
einops